static설정, 라우터 설정
