+++ public/cocktail.html
@@ -5,12 +5,12 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>칵테일 소개</title>
-    <link rel="stylesheet" href="./cocktail.css" />
+    <link rel="stylesheet" href="./stylesheets/cocktail.css" />
   </head>
   <body>
     <header>
       <div class="title-logo">
-        <img src="./img/title-logo.png" width="345px" height="150px" />
+        <img src="./images/title-logo.png" width="345px" height="150px" />
       </div>
     </header>
     <nav></nav>
@@ -36,6 +36,6 @@
         </ol>
       </div>
     </content>
-    <script src="./cocktail.js"></script>
+    <script src="./public/javascripts/cocktail.js"></script>
   </body>
 </html>
