+++ public/index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>나만의 칵테일</title>
-    <link rel="stylesheet" href="./index.css" />
+    <link rel="stylesheet" href="./stylesheets/index.css" />
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css"
       rel="stylesheet"
@@ -38,7 +38,7 @@
     <div class="body">
       <header>
         <div class="title-logo">
-          <img src="./img/title-logo.png" width="345px" height="150px" />
+          <img src="./images/title-logo.png" width="345px" height="150px" />
         </div>
         <nav>
           <button id="nav-cocktail-receipe">칵테일 레시피</button>
@@ -102,6 +102,6 @@
         </div>
       </content>
     </div>
-    <script src="./index.js"></script>
+    <script src="./javascripts/index.js"></script>
   </body>
 </html>
