+++ public/login.html
@@ -11,13 +11,13 @@
       integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65"
       crossorigin="anonymous"
     />
-    <link rel="stylesheet" href="./login.css" />
+    <link rel="stylesheet" href="./stylesheets/login.css" />
   </head>
   <body>
     <nav>
       <ul class="nav nav-pills">
         <div class="title-logo">
-          <img src="./img/title-logo.png" width="345px" height="150px" />
+          <img src="./images/title-logo.png" width="345px" height="150px" />
         </div>
         <div class="nav-right-item">
           <li class="nav-item">
@@ -47,7 +47,7 @@
       <div class="mb-3">
         <label for="pw" class="form-label">비밀번호</label>
         <input
-          type="text"
+          type="password"
           class="form-control"
           id="pw"
           placeholder="비밀번호를 입력해주세요"
@@ -64,6 +64,6 @@
       integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
       crossorigin="anonymous"
     ></script>
-    <script src="./login.js"></script>
+    <script src="./javascripts/login.js"></script>
   </body>
 </html>
