+++ public/signup.html
@@ -10,13 +10,13 @@
       integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65"
       crossorigin="anonymous"
     />
-    <link rel="stylesheet" href="./signup.css" />
+    <link rel="stylesheet" href="./stylesheets/signup.css" />
   </head>
   <body>
     <nav>
       <ul class="nav nav-pills">
         <div class="title-logo">
-          <img src="./img/title-logo.png" width="345px" height="150px" />
+          <img src="./images/title-logo.png" width="345px" height="150px" />
         </div>
         <div class="nav-right-item">
           <li class="nav-item">
@@ -87,6 +87,6 @@
       integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
       crossorigin="anonymous"
     ></script>
-    <script src="./signup.js"></script>
+    <script src="./javascripts/signup.js"></script>
   </body>
 </html>
