+++ routes/index.js
@@ -1,9 +1,10 @@
+/*
 var express = require('express');
 var router = express.Router();
 
-/* GET home page. */
 router.get('/', function(req, res, next) {
   res.render('index', { title: 'Express' });
 });
 
 module.exports = router;
+*/
