+++ routes/users.js
@@ -1,8 +1,26 @@
 var express = require('express');
 var router = express.Router();
+const mysql = require('mysql');
+
+const con = mysql.createConnection({
+  host:'localhost',
+  user: 'root',
+  password: '1234',
+  database: 'MyCocktail'
+})
+
+con.connect(function(err){
+  if(err) throw err;
+  console.log('Connected');
+})
+
+// 엔드포인트 설정
+// 아마 fetch에서 배개변수 넘겨주면 req에 들어가는듯?
+// -> "req.query.변수이름"으로 접근 가능 -> 파이썬의 딕셔너리처럼 인덱스 번호 접근이 아니라 변수 이름으로!!
+// C++에서는 아마 해시맵? 일듯?
 
 /* GET users listing. */
-router.get('/users', (req, res) => {
+router.get('/', (req, res) => {
 
   con.query("SELECT * FROM member", (err, result) =>{
     if(err){
@@ -11,6 +29,18 @@ router.get('/users', (req, res) => {
     }
     res.json(result)
   })
+})
+
+router.post('/', (req, res) => {
+  
+})
+
+router.put('/', (req, res) => {
+
+})
+
+router.delete('/', (req, res) => {
+  
 })
 
 module.exports = router;
